Assignment5/pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "onerlookedstridentattributes"
version = "0.1.0"
[tool.poetry.dependencies]
pycryptodome = "^3.9"
python = "^3.7"
simonspeckciphers = "^1.0"
[tool.poetry.dev-dependencies]
